+++ tests/frontends/mpd/reflection_test.py
@@ -1,5 +1,6 @@
 import unittest
 
+from mopidy import settings
 from mopidy.backends.dummy import DummyBackend
 from mopidy.frontends.mpd.dispatcher import MpdDispatcher
 from mopidy.mixers.dummy import DummyMixer
@@ -11,6 +12,7 @@ class ReflectionHandlerTest(unittest.TestCase):
         self.dispatcher = MpdDispatcher()
 
     def tearDown(self):
+        settings.runtime.clear()
         self.backend.stop().get()
         self.mixer.stop().get()
 
@@ -31,6 +33,19 @@ class ReflectionHandlerTest(unittest.TestCase):
         self.assert_(u'command: sticker' not in result)
         self.assert_(u'OK' in result)
 
+    def test_commands_show_less_if_auth_required_and_not_authed(self):
+        settings.MPD_SERVER_PASSWORD = u'secret'
+        result = self.dispatcher.handle_request(u'commands')
+        # Not requiring auth
+        self.assert_(u'command: close' in result, result)
+        self.assert_(u'command: commands' in result, result)
+        self.assert_(u'command: notcommands' in result, result)
+        self.assert_(u'command: password' in result, result)
+        self.assert_(u'command: ping' in result, result)
+        # Requiring auth
+        self.assert_(u'command: play' not in result, result)
+        self.assert_(u'command: status' not in result, result)
+
     def test_decoders(self):
         result = self.dispatcher.handle_request(u'decoders')
         self.assert_(u'ACK [0@0] {} Not implemented' in result)
@@ -40,6 +55,19 @@ class ReflectionHandlerTest(unittest.TestCase):
         self.assertEqual(2, len(result))
         self.assert_(u'command: kill' in result)
         self.assert_(u'OK' in result)
+
+    def test_notcommands_returns_more_if_auth_required_and_not_authed(self):
+        settings.MPD_SERVER_PASSWORD = u'secret'
+        result = self.dispatcher.handle_request(u'notcommands')
+        # Not requiring auth
+        self.assert_(u'command: close' not in result, result)
+        self.assert_(u'command: commands' not in result, result)
+        self.assert_(u'command: notcommands' not in result, result)
+        self.assert_(u'command: password' not in result, result)
+        self.assert_(u'command: ping' not in result, result)
+        # Requiring auth
+        self.assert_(u'command: play' in result, result)
+        self.assert_(u'command: status' in result, result)
 
     def test_tagtypes(self):
         result = self.dispatcher.handle_request(u'tagtypes')